requirements.txt
pandas>=2.2
pyarrow>=16.0
PyYAML>=6.0
httpx>=0.27
requests>=2.32
json5>=0.9
numpy>=2.0
psutil>=6.0
tabulate>=0.9
pytest>=8.0
